+++ liblangutil/ErrorReporter.h
@@ -23,6 +23,7 @@
 #pragma once
 
 #include <libsolutil/CommonData.h>
+#include <libsolutil/Exceptions.h>
 
 #include <liblangutil/Exceptions.h>
 #include <liblangutil/SourceLocation.h>
@@ -32,17 +33,6 @@
 
 namespace solidity::langutil
 {
-
-/**
- * Unique identifiers are used to tag and track individual error cases.
- * They are passed as the first parameter of error reporting functions.
- * Suffix _error helps to find them in the sources.
- * The struct ErrorId prevents incidental calls like typeError(3141) instead of typeError(3141_error).
- * To create a new ID, one can add 0000_error and then run "python ./scripts/correct_error_ids.py"
- * from the root of the repo.
- */
-struct ErrorId { unsigned long long error = 0; };
-ErrorId operator"" _error(unsigned long long error);
 
 class ErrorReporter
 {