+++ libsmtutil/Z3Interface.cpp
@@ -180,6 +180,19 @@ z3::expr Z3Interface::toZ3Expr(Expression const& _expr)
 			return arguments[0] / arguments[1];
 		else if (n == "mod")
 			return z3::mod(arguments[0], arguments[1]);
+		else if (n == "bvand")
+			return arguments[0] & arguments[1];
+		else if (n == "int2bv")
+		{
+			size_t size = std::stoi(_expr.arguments[1].name);
+			return z3::int2bv(size, arguments[0]);
+		}
+		else if (n == "bv2int")
+		{
+			auto intSort = dynamic_pointer_cast<IntSort>(_expr.sort);
+			smtAssert(intSort, "");
+			return z3::bv2int(arguments[0], intSort->isSigned);
+		}
 		else if (n == "select")
 			return z3::select(arguments[0], arguments[1]);
 		else if (n == "store")
@@ -194,7 +207,7 @@ z3::expr Z3Interface::toZ3Expr(Expression const& _expr)
 		}
 		else if (n == "tuple_get")
 		{
-			size_t index = std::stoi(_expr.arguments[1].name);
+			size_t index = stoul(_expr.arguments[1].name);
 			return z3::func_decl(m_context, Z3_get_tuple_sort_field_decl(m_context, z3Sort(*_expr.arguments[0].sort), index))(arguments[0]);
 		}
 		else if (n == "tuple_constructor")