+++ libsolutil/CommonData.cpp
@@ -53,6 +53,9 @@ string solidity::util::toHex(uint8_t _data, HexCase _case)
 
 string solidity::util::toHex(bytes const& _data, HexPrefix _prefix, HexCase _case)
 {
+	if (_data.empty())
+		return {};
+
 	std::string ret(_data.size() * 2 + (_prefix == HexPrefix::Add ? 2 : 0), 0);
 
 	size_t i = 0;
@@ -64,15 +67,15 @@ string solidity::util::toHex(bytes const& _data, HexPrefix _prefix, HexCase _cas
 
 	// Mixed case will be handled inside the loop.
 	char const* chars = _case == HexCase::Upper ? upperHexChars : lowerHexChars;
-	int rix = _data.size() - 1;
+	size_t rix = _data.size() - 1;
 	for (uint8_t c: _data)
 	{
 		// switch hex case every four hexchars
 		if (_case == HexCase::Mixed)
 			chars = (rix-- & 2) == 0 ? lowerHexChars : upperHexChars;
 
-		ret[i++] = chars[(unsigned(c) / 16) & 0xf];
-		ret[i++] = chars[unsigned(c) & 0xf];
+		ret[i++] = chars[(static_cast<size_t>(c) >> 4ul) & 0xfu];
+		ret[i++] = chars[c & 0xfu];
 	}
 	assertThrow(i == ret.size(), Exception, "");
 
@@ -95,6 +98,9 @@ int solidity::util::fromHex(char _i, WhenError _throw)
 
 bytes solidity::util::fromHex(std::string const& _s, WhenError _throw)
 {
+	if (_s.empty())
+		return {};
+
 	unsigned s = (_s.size() >= 2 && _s[0] == '0' && _s[1] == 'x') ? 2 : 0;
 	std::vector<uint8_t> ret;
 	ret.reserve((_s.size() - s + 1) / 2);