+++ test/TestCase.h
@@ -57,14 +57,14 @@ public:
 	/// Each line of output is prefixed with @arg _linePrefix.
 	/// If @arg _formatted is true, color-coding may be used to indicate
 	/// error locations in the contract, if applicable.
-	virtual void printSource(std::ostream &_stream, std::string const &_linePrefix = "", bool const _formatted = false) const = 0;
+	virtual void printSource(std::ostream &_stream, std::string const &_linePrefix = "", bool const _formatted = false) const;
 	/// Outputs settings.
 	virtual void printSettings(std::ostream &_stream, std::string const &_linePrefix = "", bool const _formatted = false);
 	/// Outputs updated settings
 	virtual void printUpdatedSettings(std::ostream& _stream, std::string const& _linePrefix = "");
 	/// Outputs test expectations to @arg _stream that match the actual results of the test.
 	/// Each line of output is prefixed with @arg _linePrefix.
-	virtual void printUpdatedExpectations(std::ostream& _stream, std::string const& _linePrefix) const = 0;
+	virtual void printUpdatedExpectations(std::ostream& _stream, std::string const& _linePrefix) const;
 
 	static bool isTestFilename(boost::filesystem::path const& _filename);
 
@@ -97,6 +97,11 @@ protected:
 	}
 
 	void printIndented(std::ostream& _stream, std::string const& _output, std::string const& _linePrefix = "") const;
+	TestCase::TestResult checkResult(std::ostream& _stream, const std::string& _linePrefix, bool const _formatted);
+
+	std::string m_source;
+	std::string m_obtainedResult;
+	std::string m_expectation;
 
 	TestCaseReader m_reader;
 	bool m_shouldRun = true;