+++ test/tools/yulInterpreter/Interpreter.cpp
@@ -54,7 +54,7 @@ void InterpreterState::dumpTraceAndState(ostream& _out) const
 	_out << "Memory dump:\n";
 	map<u256, u256> words;
 	for (auto const& [offset, value]: memory)
-		words[(offset / 0x20) * 0x20] |= u256(uint32_t(value)) << (256 - 8 - 8 * size_t(offset % 0x20));
+		words[(offset / 0x20) * 0x20] |= u256(uint32_t(value)) << (256 - 8 - 8 * static_cast<size_t>(offset % 0x20));
 	for (auto const& [offset, value]: words)
 		if (value != 0)
 			_out << "  " << std::uppercase << std::hex << std::setw(4) << offset << ": " << h256(value).hex() << endl;